Relocate per-kind editor logic into card modules (Phase B, #103)
Completes the Phase B split started for render logic: each kind's editor
implementation now lives in its own src/cards/<kind>.ts module alongside its
renderer, rather than being imported from the editors.ts monolith. Kind-private
editor sub-helpers move with their editor (the embed target/second-view
helpers, the calendar sources/event-note editors, the rss GitHub-feed helpers,
recentTypesEditor, refreshSetting).

editors.ts (2848 -> 383 lines) now holds only the shared settings-modal
framework — CardSettingsModal and CardSettingsOptions — plus the two generic
helpers used across kinds, addResetButton and moveItem, which the modules
import. The modal already dispatches per-kind editors through the registry
(renderEditor / editorTypeNote), so no dispatch logic changed.

Pure relocation: every moved function body is byte-identical to its previous
definition; only import blocks and the shared helpers' exports changed.
Typecheck, lint (0 errors), the production build and the full test suite (168
passing) are unchanged. README updated to describe both shared files.
